+++ Assets/Scripts/Runtime/CH1/Lucky/LuckyPMDialogue.cs
@@ -1,15 +1,15 @@
 using DG.Tweening;
 using Runtime.ETC;
-using Runtime.Luck;
+using Runtime.Lucky;
 using UnityEngine;
 using Yarn.Unity;
 
-namespace Runtime.CH1.Luck
+namespace Runtime.CH1.Lucky
 {
     public class LuckyPMDialogue : DialogueViewBase
     {
         private DialogueRunner _runner;
-        [SerializeField] private Lucky _lucky;
+        [SerializeField] private LuckyBody _lucky;
         [SerializeField] private GameObject _bubble;
         [SerializeField] private Vector3 _outPosition;
         [SerializeField] private Vector3 _inPosition;
