+++ Assets/Scripts/Runtime/CH1/Main/Player/TopDownMovement.cs
@@ -1,7 +1,7 @@
 using Runtime.Interface;
 using UnityEngine;
 
-namespace Runtime.CH1.Main.PlayerFunction
+namespace Runtime.CH1.Main.Player
 {
     public class TopDownMovement : IMovement
     {
