+++ Assets/Scripts/Runtime/CH1/Main/Player/TopDownPlayer.cs
@@ -1,4 +1,3 @@
-using Runtime.CH1.Main.PlayerFunction;
 using Runtime.ETC;
 using Runtime.Interface;
 using UnityEngine;
