+++ Assets/Scripts/Runtime/CH1/Pacmom/LuckyDialogue.cs
@@ -1,6 +1,6 @@
 using DG.Tweening;
 using Runtime.ETC;
-using Runtime.Luck;
+using Runtime.Lucky;
 using UnityEngine;
 using Yarn.Unity;
 
@@ -9,7 +9,7 @@ namespace Runtime.CH1.Pacmom
     public class LuckyDialogue : DialogueViewBase
     {
         private DialogueRunner _runner;
-        [SerializeField] private Lucky _lucky;
+        [SerializeField] private LuckyBody _lucky;
         [SerializeField] private GameObject _bubble;
         [SerializeField] private Vector3 _outPosition;
         [SerializeField] private Vector3 _inPosition;
