+++ Assets/Scripts/Runtime/CH1/SubB/SLG/SLGActionComponent.cs
@@ -3,12 +3,10 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using TMPro;
-using Runtime.CH1.Main.Stage;
 using Runtime.CH1.Main.Controller;
 using Runtime.ETC;
 using Runtime.CH1.Main.Dialogue;
-using UnityEngine.InputSystem.Controls;
-using Runtime.CH1.Luck;
+using Runtime.CH1.Lucky;
 
 namespace SLGDefines
 {
