+++ Assets/Scripts/Tests/Runtime/CH1/Player/TopDownPlayerTests.cs
@@ -1,9 +1,6 @@
 using NUnit.Framework;
-using Runtime.CH1.Main;
 using Runtime.CH1.Main.Player;
-using Runtime.CH1.Main.PlayerFunction;
 using Runtime.ETC;
-using Runtime.InGameSystem;
 using Runtime.Interface;
 using System.Collections;
 using UnityEditor.Animations;
